refactor: 가용 인원 패널 통일 + 매트릭스 사이드바 순서 교체
#1 매트릭스 SidePanel(풍부한 가용 인원 카드) 을 MemberAvailabilityPanel
   공용 컴포넌트로 추출. 주차별 UI / 멤버·곡 탭 사이드바도 동일 컴포넌트 사용
   - components/MemberAvailabilityPanel.client.tsx 신설 (헤더/진행도/가능·불가/제거 버튼)
   - HoveredAvailabilityPanel 은 store 구독 + buildCoverageHeatmap 합성 후
     MemberAvailabilityPanel 로 위임하는 wrapper 로 단순화
   - 매트릭스 내부 SidePanel/MemberChip 정의 제거

#2 매트릭스 우측 사이드바 순서 교체
   - 이전: 가용 인원(상) + 합주 블럭 풀(하)
   - 변경: 합주 블럭 풀(상) + 가용 인원(하)
