+++ src/domain/schedule-coordination/components/MatrixView.client.tsx
@@ -19,6 +19,7 @@ import { copyWeekBlocks } from '../utils/copyWeekBlocks';
 
 import type { Member } from '@/domain/setlist-meeting/types';
 
+import { MemberAvailabilityPanel } from './MemberAvailabilityPanel.client';
 import { songTone } from './palette';
 import { RANGE_PRESETS, DEFAULT_RANGE_PRESET } from './rangePresets';
 import { ScheduleBlockPanel } from './ScheduleBlockPanel.client';
@@ -671,9 +672,17 @@ export function MatrixView({
       </div>
 
       {/* 우측 사이드바 — 마스터(가능/불가) + 슬레이브(블록 풀) 세로 stack. */}
+      {/* 합주 블럭 풀(상) + 가용 인원(하) — 사용자 요청에 따라 순서 변경. */}
       <aside className="flex w-60 shrink flex-col gap-4 lg:w-72 xl:w-80">
-        <SidePanel
-          focus={focus}
+        <BlockPoolPanel
+          songPool={songPool}
+          onDragStart={onPoolDragStart}
+          onDragEnd={onPoolDragEnd}
+          focusedSongId={focusedSongId}
+          onSongClick={(songId) => toggleFocusedSongId(meetingId, songId)}
+        />
+        <MemberAvailabilityPanel
+          cell={focus}
           pinned={!!pinned}
           availability={availability}
           participants={participants}
@@ -682,13 +691,7 @@ export function MatrixView({
           block={focus ? (lockAt(focus.date, focus.slot) ?? null) : null}
           onUnpin={() => setPinned(null)}
           onRemoveBlock={(blockId) => boardId && removeBlock(boardId, blockId)}
-        />
-        <BlockPoolPanel
-          songPool={songPool}
-          onDragStart={onPoolDragStart}
-          onDragEnd={onPoolDragEnd}
-          focusedSongId={focusedSongId}
-          onSongClick={(songId) => toggleFocusedSongId(meetingId, songId)}
+          placeholder={'매트릭스 셀을 클릭하면\n해당 시간의 멤버 가용성이 표시됩니다.'}
         />
       </aside>
 
@@ -889,165 +892,6 @@ function ConfirmPreviewCard({
   );
 }
 
-function SidePanel({
-  focus,
-  pinned,
-  availability,
-  participants,
-  memberSchedules,
-  songMap,
-  block,
-  onUnpin,
-  onRemoveBlock,
-}: {
-  focus: CellRef | null;
-  pinned: boolean;
-  availability: Map<string, Set<string>>;
-  participants: Member[];
-  memberSchedules: MemberSchedule[];
-  songMap: Map<string, SongLite>;
-  block: ScheduleBlock | null;
-  onUnpin: () => void;
-  onRemoveBlock: (blockId: string) => void;
-}) {
-  if (!focus) {
-    return (
-      <div className="bg-card border-border flex flex-1 flex-col items-center justify-center gap-2 rounded-md border p-6 text-center shadow-sm">
-        <Info className="text-foreground-muted/60 h-5 w-5" />
-        <p className="text-foreground-muted text-caption">
-          매트릭스의 셀에 호버하면
-          <br />
-          해당 시간의 멤버 가용성이 표시됩니다.
-        </p>
-      </div>
-    );
-  }
-
-  const dow = dayOfWeek(focus.date);
-  const dowLabel = DOW_LABEL[dow];
-  const availableIds = availability.get(`${focus.date}__${focus.slot}`) ?? new Set<string>();
-  const total = participants.length;
-
-  const unavailable: Array<{ member: Member; reason?: string }> = [];
-  const available: Member[] = [];
-  for (const m of participants) {
-    if (availableIds.has(m.id)) {
-      available.push(m);
-    } else {
-      const sched = memberSchedules.find((s) => s.userId === m.id);
-      const reason = !sched
-        ? '미입력'
-        : !sched.availableDates.includes(focus.date)
-          ? '선약'
-          : '시간';
-      unavailable.push({ member: m, reason });
-    }
-  }
-
-  const blockedSong = block?.songId ? songMap.get(block.songId) : null;
-
-  return (
-    <div className="bg-card border-border flex flex-1 flex-col overflow-hidden rounded-md border shadow-sm">
-      <div className="border-border px-s-4 py-s-3 border-b">
-        <div className="text-foreground-muted text-micro gap-s-1 flex items-center font-bold uppercase">
-          {block ? (
-            <>
-              <Lock className="h-3 w-3" /> 합주 배치됨
-            </>
-          ) : pinned ? (
-            <>
-              <Pin className="h-3 w-3" /> 고정됨
-            </>
-          ) : (
-            <>선택됨</>
-          )}
-        </div>
-        <div className="text-foreground text-body mt-0.5 font-mono font-bold">
-          {focus.date} ({dowLabel}) {slotToTime(focus.slot)}
-        </div>
-        {blockedSong && (
-          <div className="text-accent text-caption mt-0.5 font-bold">{blockedSong.title}</div>
-        )}
-        <div className="mt-s-2 gap-s-2 flex items-center">
-          <div
-            className="bg-bg/40 h-2 flex-1 overflow-hidden rounded-full"
-            role="progressbar"
-            aria-valuenow={available.length}
-            aria-valuemin={0}
-            aria-valuemax={total}
-          >
-            <div
-              className={cn(
-                'h-full transition-all',
-                available.length === total ? 'bg-success' : 'bg-accent',
-              )}
-              style={{
-                width: `${total === 0 ? 0 : Math.round((available.length / total) * 100)}%`,
-              }}
-            />
-          </div>
-          <span className="text-caption text-foreground font-bold tabular-nums">
-            {available.length}/{total}
-          </span>
-        </div>
-        {pinned && !block && (
-          <button
-            type="button"
-            onClick={onUnpin}
-            className="text-foreground-muted hover:text-foreground text-micro mt-s-2 inline-flex items-center gap-1"
-          >
-            <X className="h-3 w-3" /> 고정 해제
-          </button>
-        )}
-      </div>
-
-      <div className="flex-1 overflow-y-auto">
-        {available.length > 0 && (
-          <div className="px-s-3 py-s-3">
-            <div className="text-success text-caption mb-s-2 gap-s-1 flex items-center font-bold">
-              <Check className="h-3.5 w-3.5" /> 가능 ({available.length}명)
-            </div>
-            <ul className="gap-s-1 flex flex-col">
-              {available.map((m) => (
-                <li key={m.id}>
-                  <MemberChip member={m} variant="ok" />
-                </li>
-              ))}
-            </ul>
-          </div>
-        )}
-        {unavailable.length > 0 && (
-          <div className="border-border px-s-3 py-s-3 border-t">
-            <div className="text-danger text-caption mb-s-2 gap-s-1 flex items-center font-bold">
-              <X className="h-3.5 w-3.5" /> 불가 ({unavailable.length}명)
-            </div>
-            <ul className="gap-s-1 flex flex-col">
-              {unavailable.map(({ member, reason }) => (
-                <li key={member.id}>
-                  <MemberChip member={member} variant="no" reason={reason} />
-                </li>
-              ))}
-            </ul>
-          </div>
-        )}
-      </div>
-
-      {block && (
-        <div className="border-border px-s-3 py-s-3 border-t">
-          <Button
-            variant="ghost"
-            size="sm"
-            onClick={() => onRemoveBlock(block.blockId)}
-            className="text-danger w-full"
-          >
-            <Unlock className="h-4 w-4" /> 이 합주 슬롯 제거
-          </Button>
-        </div>
-      )}
-    </div>
-  );
-}
-
 function BlockPoolPanel({
   songPool,
   onDragStart,
@@ -1105,40 +949,6 @@ function BlockPoolPanel({
           })
         )}
       </div>
-    </div>
-  );
-}
-
-function MemberChip({
-  member,
-  variant,
-  reason,
-}: {
-  member: Member;
-  variant: 'ok' | 'no';
-  reason?: string;
-}) {
-  const initial = member.name.slice(0, 1);
-  return (
-    <div
-      className={cn(
-        'gap-s-2 flex items-center rounded-md px-2 py-1.5',
-        variant === 'ok' ? 'bg-success-dim' : 'bg-danger-dim grayscale-[0.4]',
-      )}
-    >
-      <span
-        className="text-bg inline-flex h-6 w-6 shrink-0 items-center justify-center rounded-full text-xs font-bold"
-        style={{ backgroundColor: member.avatar ?? '#6b6b80' }}
-      >
-        {initial}
-      </span>
-      <span className="text-caption flex-1 truncate font-bold">{member.name}</span>
-      <span className="text-foreground-muted text-micro">{member.role ?? ''}</span>
-      {variant === 'no' && reason && (
-        <span className="bg-danger-dim text-danger text-micro px-s-2 rounded-full py-0.5 font-bold">
-          {reason}
-        </span>
-      )}
     </div>
   );
 }
